adversarial fiddling

+++ Code/Models/Attention_seq2seq.py
@@ -387,8 +387,8 @@ class _Seq2Seq(nn.Module):
         # introduce noise if adversarial = True
         if adversarial == True:
             hidden = hidden_.clone() + torch.from_numpy(
-                np.random.normal(0, 0.01, size = hidden_.shape), dtype = hidden_.dtype
-                ).to(self.device)
+                np.random.normal(0, 0.01, size = hidden_.shape)
+                ).double().to(self.device)
         else:
             hidden = hidden_
